Move core/layouts to shared/layouts.
