+++ src/app/app.module.ts
@@ -7,8 +7,8 @@ import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { CoreModule } from './core/core.module';
 import { AuthInterceptor } from './core/interceptors/auth.interceptor';
-import { AuthLayoutComponent } from './core/layouts/auth-layout/auth-layout.component';
-import { MainLayoutComponent } from './core/layouts/main-layout/main-layout.component';
+import { AuthLayoutComponent } from './shared/layouts/auth-layout/auth-layout.component';
+import { MainLayoutComponent } from './shared/layouts/main-layout/main-layout.component';
 import { AuthService } from './core/services/auth.service';
 import { MainAppModule } from './modules/main-app.module';
 import { PageNotFoundComponent } from './shared/pages/page-not-found/page-not-found.component';
