+++ src/app/core/core.module.ts
@@ -1,8 +1,8 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { AuthModule } from './auth/auth.module';
-import { AuthLayoutComponent } from './layouts/auth-layout/auth-layout.component';
-import { MainLayoutComponent } from './layouts/main-layout/main-layout.component';
+import { AuthLayoutComponent } from '../shared/layouts/auth-layout/auth-layout.component';
+import { MainLayoutComponent } from '../shared/layouts/main-layout/main-layout.component';
 
 @NgModule({
   declarations: [],
