+++ src/app/shared/layouts/auth-layout/auth-layout.component.html
